+++ py/sight/demo/agentic_demo/bayes_opt_llm_demo.py
