+++ py/sight/demo/agentic_demo/bayes_opt_worker.py
