+++ py/sight/demo/agentic_demo/calculator_demo.py
@@ -19,5 +19,5 @@ agent = initialize_agent(
     verbose=True,
 )
 
-response = agent.invoke({"input": "can you divide 200 to 10?"})
+response = agent.invoke({"input": "can you divide 200 to 10 and add 20 to the result"})
 print(response)
