+++ py/sight/tests/extra/test_launch_worker.py
