+++ py/sight/tests/extra/test_rpc.py
