+++ py/sight/utils/text_proto_generator.py
