+++ py/worker/.text_proto_configs/calculator.textproto
