+++ py/worker/.text_proto_configs/generic.textproto
