+++ py/worker/worker_configs/calculator_worker.yaml
@@ -1,3 +1,3 @@
 v0.1 :
   docker: "gcr.io/cameltrain/test-calculator-worker:latest"
-  binary: "worker/binary/calculator_worker.py"
+  binary: "py/worker/binary/calculator_worker.py"
