+++ py/worker/worker_configs/fvs_worker.yaml
@@ -1,3 +1,3 @@
 v0.1 :
   docker: "gcr.io/cameltrain/test-generic-worker-local:latest"
-  binary: "worker/binary/fvs_worker.py"
+  binary: "py/worker/binary/fvs_worker.py"
