+++ py/worker/worker_configs/generic_worker.yaml
@@ -1,3 +1,3 @@
 v0.1 :
   docker: "gcr.io/cameltrain/test-generic-worker-local:latest"
-  binary: "py/sight/demo/w1.py"
+  binary: "py/worker/demo/w1.py"
